worker/requirements.txt
redis==4.5.5
pymongo==4.3.3
python-dotenv==1.0.0
